src/backend/requirements.txt
Flask==2.0.2
requests==2.26.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.2
python-dotenv==0.19.2
flask-swagger-ui==3.36.0
vowpalwabbit==8.11.0
pandas==1.2.4
matplotlib==3.5.0
pytest==6.2.5
pycamunda==0.5.0
streamlit==1.3.1